Final prep before merge
Divided admin file into multiple file per route
Moved utils into a folder
Added more tests on utils
