+++ openapprentice/utils/url.py
@@ -1,0 +1,42 @@
+# -*- coding: utf-8 -*-
+
+"""
+    The OpenApprentice Foundation and its website OpenApprentice.org
+    Copyright (C) 2018 The OpenApprentice Foundation - contact@openapprentice.org
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+This file contains all utility function, classes and variables used throughout the program
+"""
+
+import sys
+
+from flask import request
+
+if sys.version_info > (3, 0):
+    from urllib.parse import urlparse, urljoin
+else:
+    from urlparse import urlparse, urljoin
+
+
+def is_safe_url(target):
+    """
+    Tests if the url to redirect to is safe
+    :param target: the url to redirect to
+    :return: True if the url is safe, False otherwise
+    """
+
+    ref_url = urlparse(request.host_url)
+    test_url = urlparse(urljoin(request.host_url, target))
+    return test_url.scheme in ('http', 'https') and ref_url.netloc == test_url.netloc
